requirements.txt
# Core data
numpy>=1.26
pandas>=2.1
pyarrow>=14.0

# Database
psycopg2-binary>=2.9

# ML
scikit-learn>=1.4
xgboost>=2.0
lightgbm>=4.2
shap>=0.44

# Visualization & Dashboard
plotly>=5.18
dash>=2.14
dash-bootstrap-components>=1.5
matplotlib>=3.8
kaleido>=0.2

# Reporting
jinja2>=3.1
reportlab>=4.0
weasyprint>=60.0

# Utilities
pyyaml>=6.0
click>=8.1
joblib>=1.3
tqdm>=4.66

# Dev / Testing
pytest>=7.4
ruff>=0.1
